change files for new robots
