+++ botsim/public/bots/knotechcallibot/chassis.png
