+++ mkc-knotechcallibot.json
@@ -1,0 +1,5 @@
+{
+    "overrides": {
+        "testFiles": ["remotes/mainknotechcallibot.ts"]
+    }
+}
