--- mkc-knotechcallibot1.json
@@ -1,5 +1,0 @@
-{
-    "overrides": {
-        "testFiles": ["remotes/mainknotechcallibot1.ts"]
-    }
-}
